Moved sample spectral plots to misc.py.
