+++ colorpy/plots.py
@@ -91,12 +91,6 @@ color_vs_param_plot (
 
 Specialized plots:
 
-visible_spectrum_plot () -
-    Plot the visible spectrum, as a plot vs wavelength.
-
-cie_matching_functions_plot () -
-    Plot the CIE XYZ matching functions, as three spectral subplots.
-
 shark_fin_plot () -
     Draw the 'shark fin' CIE chromaticity diagram of the pure spectral lines (plus purples) in xy space.
 
@@ -114,7 +108,6 @@ import matplotlib.pyplot as pyplot
 import colormodels
 import ciexyz
 import pure_colors
-import rayleigh
 
 #
 # Utilities for plots.
@@ -373,119 +366,6 @@ def color_vs_param_plot (
 # Some specialized figures.
 #
 
-def visible_spectrum_plot (filename='VisibleSpectrum'):
-    ''' Plot the visible spectrum, as a plot vs wavelength. '''
-    spect = ciexyz.Spectrum()
-    # Get rgb colors for each wavelength.
-    rgb_colors = numpy.empty ((spect.num_wl, 3))
-    for i in range (spect.num_wl):
-        xyz = ciexyz.xyz_from_wavelength (spect.wavelength [i])
-        rgb = colormodels.rgb_from_xyz (xyz)
-        rgb_colors [i] = rgb
-    # Scale to make brightest rgb value = 1.0.
-    rgb_max = numpy.max (rgb_colors)
-    if rgb_max != 0.0:
-        scaling = 1.0 / rgb_max
-        rgb_colors *= scaling
-    # plot colors and rgb values vs wavelength
-    color_vs_param_plot (
-        spect.wavelength,
-        rgb_colors,
-        'The Visible Spectrum',
-        filename,
-        tight = True,
-        xlabel = r'Wavelength (nm)',
-        ylabel = r'RGB Color')
-
-
-def cie_matching_functions_plot ():
-    ''' Plot the CIE XYZ matching functions, as three spectral subplots. '''
-    # Get 'spectra' for x,y,z matching functions.
-    spect_x = ciexyz.Spectrum()
-    spect_y = ciexyz.Spectrum()
-    spect_z = ciexyz.Spectrum()
-    for i in range (spect_x.num_wl):
-        wl_nm = spect_x.wavelength [i]
-        xyz = ciexyz.xyz_from_wavelength (wl_nm)
-        spect_x.intensity [i] = xyz [0]
-        spect_y.intensity [i] = xyz [1]
-        spect_z.intensity [i] = xyz [2]
-    # Plot three separate subplots, with CIE X in the first,
-    # CIE Y in the second, and CIE Z in the third.
-    # Label appropriately for the whole plot.
-    pyplot.clf ()
-    # X
-    pyplot.subplot (3,1,1)
-    pyplot.title ('1931 CIE XYZ Matching Functions')
-    pyplot.ylabel ('CIE $X$')
-    spectrum_subplot (spect_x)
-    tighten_x_axis (spect_x.wavelength)
-    # Y
-    pyplot.subplot (3,1,2)
-    pyplot.ylabel ('CIE $Y$')
-    spectrum_subplot (spect_y)
-    tighten_x_axis (spect_y.wavelength)
-    # Z
-    pyplot.subplot (3,1,3)
-    pyplot.xlabel ('Wavelength (nm)')
-    pyplot.ylabel ('CIE $Z$')
-    spectrum_subplot (spect_z)
-    tighten_x_axis (spect_z.wavelength)
-    # Save.
-    filename = 'CIEXYZ_Matching'
-    plot_save (filename)
-
-
-def cie_matching_functions_spectrum_plot ():
-    ''' Plot each of the CIE XYZ matching functions, as spectrum plots. '''
-    # Get 'spectra' for x,y,z matching functions.
-    spect_x = ciexyz.Spectrum()
-    spect_y = ciexyz.Spectrum()
-    spect_z = ciexyz.Spectrum()
-    for i in range (spect_x.num_wl):
-        wl_nm = spect_x.wavelength [i]
-        xyz = ciexyz.xyz_from_wavelength (wl_nm)
-        spect_x.intensity [i] = xyz [0]
-        spect_y.intensity [i] = xyz [1]
-        spect_z.intensity [i] = xyz [2]
-    # Create three spectrum plots.
-    spectrum_plot (spect_x, 'CIE X', 'CIE-X')
-    spectrum_plot (spect_y, 'CIE Y', 'CIE-Y')
-    spectrum_plot (spect_z, 'CIE Z', 'CIE-Z')
-
-
-def scattered_visual_brightness ():
-    ''' Plot the perceptual brightness of Rayleigh scattered light. '''
-    # This combines the extent of scattering with how bright it appears.
-    # It shows why a green laser shows a distinct trail in the air,
-    # while a red laser does not.
-    # Get 'spectra' as CIE Y matching function and multiply by scattering.
-    # Rayleigh scattering is proportional to 1 / wl^4.
-    spect = ciexyz.Spectrum()
-    for i in range (spect.num_wl):
-        wl_nm = spect.wavelength [i]
-        xyz = ciexyz.xyz_from_wavelength (wl_nm)
-        scatter = rayleigh.rayleigh_scattering (wl_nm)
-        spect.intensity [i] = xyz[1] * scatter
-    # Scale is arbitrary so make max intensity nearly 1.0.
-    # It looks a little better if the max is just under 1.
-    max_intensity = max (spect.intensity)
-    want_max = 0.99
-    if max_intensity != 0.0:
-        scaling = want_max / max_intensity
-        spect.intensity *= scaling
-    # Plot.
-    pyplot.clf ()
-    pyplot.title ('Perceptual Brightness of Rayleigh Scattered Light')
-    pyplot.xlabel ('Wavelength (nm)')
-    pyplot.ylabel ('CIE $Y$ / $\lambda^4$')
-    spectrum_subplot (spect)
-    tighten_x_axis (spect.wavelength)
-    # Save.
-    filename = 'Laser-Scatter'
-    plot_save (filename)
-
-
 def shark_fin_plot ():
     '''Draw the 'shark fin' CIE chromaticity diagram of the pure spectral lines (plus purples) in xy space.'''
     # get array of (approximate) colors for the boundary of the fin
@@ -602,11 +482,7 @@ def shark_fin_plot ():
 
 def figures ():
     '''Draw specific figures not used anywhere else.'''
-    visible_spectrum_plot()
-    cie_matching_functions_plot()
-    cie_matching_functions_spectrum_plot()
     shark_fin_plot()
-    scattered_visual_brightness()
 
 
 if __name__ == '__main__':
